# openespi-authserver/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenESPI Authorization Server API
  description: |
    OAuth2/OIDC Authorization Server for Green Button Alliance ESPI 4.0 compliant applications.
    
    This API provides:
    - OAuth2 client management
    - DataCustodian integration
    - OIDC UserInfo endpoint with ESPI claims
    - NAESB ESPI 4.0 compliance features
    
    ## Security
    
    - **TLS 1.3 ONLY**: All communications must use TLS 1.3
    - **Certificate Authentication**: Support for X.509 client certificates
    - **Bearer Token**: Required for API access
    - **Perfect Forward Secrecy**: All cipher suites support PFS
    
    ## ESPI Compliance
    
    This server is certified for NAESB ESPI 4.0 compliance and supports:
    - Green Button Connect My Data (CMD)
    - Function Block 4.5.15+ 
    - Certificate-based client authentication
    - ESPI-specific scope validation
  version: 1.0.0
  contact:
    name: Green Button Alliance
    email: support@greenbuttonalliance.org
    url: https://www.greenbuttonalliance.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  termsOfService: https://www.greenbuttonalliance.org/terms
servers:
  - url: https://authorization.greenbuttonalliance.org
    description: Production server
  - url: https://staging-authorization.greenbuttonalliance.org
    description: Staging server
security:
  - BearerAuth: []
  - ClientCertificate: []
paths:
...
  /api/v1/datacustodian/verify-user:
    post:
      tags:
        - DataCustodian Integration
      summary: Verify user credentials
      description: Verify user credentials with DataCustodian
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyUserRequest'
      responses:
        '200':
          description: User verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/DataCustodianError'
  /api/v1/datacustodian/customers/{customerId}:
    get:
      tags:
        - DataCustodian Integration
      summary: Get retail customer info
      description: Retrieve retail customer information from DataCustodian
      security:
        - BearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          description: DataCustodian customer identifier
          schema:
            type: string
      responses:
        '200':
          description: Customer information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetailCustomerResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/DataCustodianError'
  /api/v1/datacustodian/customers/{customerId}/usage-points:
    get:
      tags:
        - DataCustodian Integration
      summary: Get customer usage points
      description: Retrieve usage points for a customer
      security:
        - BearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          description: DataCustodian customer identifier
          schema:
            type: string
      responses:
        '200':
          description: Usage points information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsagePointsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/DataCustodianError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ClientCredentials:
      type: http
      scheme: basic
    ClientCertificate:
      type: mutualTLS
  schemas:
...
    VerifyUserRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    VerifyUserResponse:
      type: object
      properties:
        verified:
          type: boolean
        customerId:
          type: string
        customerType:
          type: string
          enum: [RESIDENTIAL, COMMERCIAL, INDUSTRIAL]
        message:
          type: string
    RetailCustomerResponse:
      type: object
      properties:
        customerId:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        customerType:
          type: string
          enum: [RESIDENTIAL, COMMERCIAL, INDUSTRIAL]
        accountNumber:
          type: string
        serviceTerritory:
          type: string
        enrollmentDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
    UsagePointsResponse:
      type: object
      properties:
        customerId:
          type: string
        usagePoints:
          type: array
          items:
            $ref: '#/components/schemas/UsagePoint'
    UsagePoint:
      type: object
      properties:
        usagePointId:
          type: string
        usagePointUUID:
          type: string
          format: uuid
        serviceCategory:
          type: string
          enum: [ELECTRICITY, GAS, WATER, TIME]
        serviceKind:
          type: string
          enum: [ENERGY, DEMAND, VOLTAGE, CURRENT]
        meterNumber:
          type: string
        serviceAddress:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        installationDate:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        error_description:
          type: string
        error_uri:
          type: string
          format: uri
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        correlationId:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "invalid_request"
            error_description: "The request is malformed"
            timestamp: "2024-01-16T15:30:00Z"
            path: "/api/v1/oauth2/clients"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "invalid_token"
            error_description: "The access token is invalid or expired"
            timestamp: "2024-01-16T15:30:00Z"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "client_not_found"
            error_description: "The specified client was not found"
            timestamp: "2024-01-16T15:30:00Z"
...
    DataCustodianError:
      description: DataCustodian integration error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "datacustodian_error"
            error_description: "DataCustodian service is unavailable"
            timestamp: "2024-01-16T15:30:00Z"
...